cambiando tipos de dato de double a float

+++ factura.groovy
@@ -3,7 +3,7 @@ class Facturacion{ Factura
 	Emisor emisor
 	Cliente cliente
 	List<Concepto> concepto = []
-	Double subtotal
-	Double iva
-	Double total
+	Float subtotal
+	Float iva
+	Float total
 }
